pyproject.toml
[tool.poetry]
name = "repl_python3_rsvpbot-1"
version = "0.1.0"
description = ""
authors = ["kwoo99 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
